# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: M&E Garments storefront and shopping-assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storefront/home:
    get:
      operationId: getStorefrontHome
      tags: [storefront]
      summary: Get real Shopify-powered homepage merchandising
      responses:
        "200":
          description: Homepage catalog content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorefrontHome"
        "503":
          description: Shopify is not connected or temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /storefront/search:
    get:
      operationId: searchProducts
      tags: [storefront]
      summary: Search the live Shopify catalog
      parameters:
        - name: query
          in: query
          schema: { type: string, maxLength: 200 }
        - name: collection
          in: query
          schema: { type: string }
        - name: age
          in: query
          schema: { type: string }
        - name: occasion
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Matching real products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSearchResult"
        "503":
          description: Shopify is not connected or temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /storefront/products/{handle}:
    get:
      operationId: getProduct
      tags: [storefront]
      summary: Get a live Shopify product by handle
      parameters:
        - name: handle
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
      responses:
        "200":
          description: Product detail from Shopify
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is not connected or temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /storefront/cart:
    get:
      operationId: getCart
      tags: [storefront]
      summary: Get a Shopify cart
      parameters:
        - name: cartId
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Current Shopify cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "404":
          description: Cart not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createCart
      tags: [storefront]
      summary: Create a Shopify cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartCreateInput"
      responses:
        "201":
          description: Created Shopify cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid cart line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /storefront/cart/lines:
    post:
      operationId: addCartLine
      tags: [storefront]
      summary: Add a verified variant to a Shopify cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartLineAddInput"
      responses:
        "200":
          description: Updated Shopify cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid cart line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateCartLine
      tags: [storefront]
      summary: Update a Shopify cart line quantity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartLineUpdateInput"
      responses:
        "200":
          description: Updated Shopify cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid cart line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /storefront/cart/lines/remove:
    post:
      operationId: removeCartLine
      tags: [storefront]
      summary: Remove a line from a Shopify cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartLineRemoveInput"
      responses:
        "200":
          description: Updated Shopify cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Invalid cart line
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Shopify is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /assistant/query:
    post:
      operationId: queryShoppingAssistant
      tags: [assistant]
      summary: Ask the grounded M&E shopping assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQueryInput"
      responses:
        "200":
          description: Grounded assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: AI or Shopify is not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
      required: [error, code]
...
    Money:
      type: object
      properties:
        amount: { type: string }
        currencyCode: { type: string }
      required: [amount, currencyCode]
    ProductImage:
      type: object
      properties:
        url: { type: string, format: uri }
        altText: { type: ["string", "null"] }
        width: { type: ["integer", "null"] }
        height: { type: ["integer", "null"] }
      required: [url, altText, width, height]
    Product:
      type: object
      properties:
        id: { type: string }
        handle: { type: string }
        title: { type: string }
        description: { type: string }
        image:
          oneOf:
            - $ref: "#/components/schemas/ProductImage"
            - type: "null"
        price:
          $ref: "#/components/schemas/Money"
        compareAtPrice:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        availableForSale: { type: boolean }
        sizes:
          type: array
          items: { type: string }
        badges:
          type: array
          items: { type: string }
      required: [id, handle, title, description, image, price, compareAtPrice, availableForSale, sizes, badges]
    SelectedOption:
      type: object
      properties:
        name: { type: string }
        value: { type: string }
      required: [name, value]
    ProductOption:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        values:
          type: array
          items: { type: string }
      required: [id, name, values]
    ProductVariant:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        availableForSale: { type: boolean }
        quantityAvailable: { type: ["integer", "null"] }
        price:
          $ref: "#/components/schemas/Money"
        compareAtPrice:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        image:
          oneOf:
            - $ref: "#/components/schemas/ProductImage"
            - type: "null"
        selectedOptions:
          type: array
          items:
            $ref: "#/components/schemas/SelectedOption"
      required: [id, title, availableForSale, quantityAvailable, price, compareAtPrice, image, selectedOptions]
    ProductDetail:
      type: object
      properties:
        id: { type: string }
        handle: { type: string }
        title: { type: string }
        description: { type: string }
        descriptionHtml: { type: string }
        vendor: { type: string }
        productType: { type: string }
        availableForSale: { type: boolean }
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        options:
          type: array
          items:
            $ref: "#/components/schemas/ProductOption"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
      required: [id, handle, title, description, descriptionHtml, vendor, productType, availableForSale, images, options, variants]
    CartLineProduct:
      type: object
      properties:
        handle: { type: string }
        title: { type: string }
      required: [handle, title]
    CartLineMerchandise:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        price:
          $ref: "#/components/schemas/Money"
        compareAtPrice:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        image:
          oneOf:
            - $ref: "#/components/schemas/ProductImage"
            - type: "null"
        selectedOptions:
          type: array
          items:
            $ref: "#/components/schemas/SelectedOption"
        product:
          $ref: "#/components/schemas/CartLineProduct"
      required: [id, title, price, compareAtPrice, image, selectedOptions, product]
    CartLine:
      type: object
      properties:
        id: { type: string }
        quantity: { type: integer, minimum: 1 }
        merchandise:
          $ref: "#/components/schemas/CartLineMerchandise"
      required: [id, quantity, merchandise]
    CartCost:
      type: object
      properties:
        subtotalAmount:
          $ref: "#/components/schemas/Money"
        totalAmount:
          $ref: "#/components/schemas/Money"
      required: [subtotalAmount, totalAmount]
    Cart:
      type: object
      properties:
        id: { type: string }
        checkoutUrl: { type: string, format: uri }
        totalQuantity: { type: integer, minimum: 0 }
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
        cost:
          $ref: "#/components/schemas/CartCost"
      required: [id, checkoutUrl, totalQuantity, lines, cost]
    CartCreateInput:
      type: object
      properties:
        merchandiseId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 20, default: 1 }
      required: [merchandiseId]
    CartLineAddInput:
      type: object
      properties:
        cartId: { type: string, minLength: 1 }
        merchandiseId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 20, default: 1 }
      required: [cartId, merchandiseId]
    CartLineUpdateInput:
      type: object
      properties:
        cartId: { type: string, minLength: 1 }
        lineId: { type: string, minLength: 1 }
        quantity: { type: integer, minimum: 1, maximum: 20 }
      required: [cartId, lineId, quantity]
    CartLineRemoveInput:
      type: object
      properties:
        cartId: { type: string, minLength: 1 }
        lineId: { type: string, minLength: 1 }
      required: [cartId, lineId]
    MerchandisingGroup:
      type: object
      properties:
        title: { type: string }
        handle: { type: string }
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [title, handle, products]
    StorefrontHome:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/MerchandisingGroup"
      required: [groups]
    ProductSearchResult:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total: { type: integer, minimum: 0 }
      required: [products, total]
    AssistantQueryInput:
      type: object
      properties:
        message: { type: string, minLength: 1, maxLength: 1000 }
        conversationId: { type: ["string", "null"] }
      required: [message]
    AssistantResponse:
      type: object
      properties:
        conversationId: { type: string }
        message: { type: string }
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
      required: [conversationId, message, products]