# lib/api-zod/src/generated/types/cartLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { CartLineMerchandise } from './cartLineMerchandise';

export interface CartLine {
  id: string;
  /** @minimum 1 */
  quantity: number;
  merchandise: CartLineMerchandise;
}