# lib/api-zod/src/generated/types/cartLineAddInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CartLineAddInput {
  /** @minLength 1 */
  cartId: string;
  /** @minLength 1 */
  merchandiseId: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  quantity?: number;
}