# lib/api-zod/src/generated/types/cartLineProduct.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CartLineProduct {
  handle: string;
  title: string;
}