# lib/api-zod/src/generated/types/storefrontHome.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * M&E Garments storefront and shopping-assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { MerchandisingGroup } from './merchandisingGroup';

export interface StorefrontHome {
  groups: MerchandisingGroup[];
}